+++ lib/src/data/global_widgets/wish_icon.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:food_app/src/controller/item_provider.dart';
+import 'package:food_app/src/controller/wish_list_controller.dart';
 import 'package:provider/provider.dart';
 
 class WishIcon extends StatefulWidget {
@@ -13,17 +14,17 @@ class WishIcon extends StatefulWidget {
 class _WishIconState extends State<WishIcon> {
   @override
   Widget build(BuildContext context) {
-    var itemProvider = Provider.of<ItemProvider>(context);
+    var wishLIstProvider = Provider.of<WishListController>(context);
     return InkWell(
         onTap: () {
-          itemProvider.setIsFavorite();
+
         },
         child: Container(
           height: 40.h,
           width: 40.w,
           decoration:
               const BoxDecoration(color: Colors.white, shape: BoxShape.circle),
-          child: itemProvider.isFavorite
+          child: wishLIstProvider.isFavorite
               ? const Icon(
                   Icons.favorite_border,
                   color: Colors.red,
