+++ lib/src/modules/search/search_screen.dart
@@ -1,14 +1,19 @@
 import 'package:flutter/material.dart';
+import 'package:food_app/src/controller/wish_list_controller.dart';
+import 'package:provider/provider.dart';
 
 class SearchScreen extends StatelessWidget {
   const SearchScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
+    //var wishListProvider = Provider.of<WishListController>(context);
     return Scaffold(
-      body: Center(
-        child: Image.asset('assets/images/dish_two.png')
-      ),
+      body: ListView.builder(
+        itemCount: 5,
+          itemBuilder: (context, index){
+          return Text(index.toString());
+      })
     );
   }
 }
